+++ CHANGELOG.md
@@ -527,3 +527,4 @@
 - [2026-04-18 15:41] feat(forms): implement responsive forms view component
 - [2026-04-18 15:48] fix(auth): fix memory leak by cancelling auth subscription
 - [2026-04-18 16:32] perf(auth): apply RepaintBoundary on animated auth widgets
+- [2026-04-18 16:55] feat(i18n): add async i18n stream provider